Move codegen_ssa SIR bits into their own file.
I'd have liked for this stuff to have gone into the SirFuncCx, but we
can't do that, as backend types can't appear in rustc_middle.
