+++ src/librustc_codegen_ssa/lib.rs
@@ -40,6 +40,7 @@ pub mod glue;
 pub mod meth;
 pub mod mir;
 pub mod mono_item;
+pub mod sir;
 pub mod traits;
 
 pub struct ModuleCodegen<M> {
